@@ -20,17 +20,17 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "illuminate/contracts": "~5.5|~6|~7|~8|~9",
-        "illuminate/support": "~5.5|~6|~7|~8|~9",
+        "php": "^8.1",
+        "illuminate/contracts": "~8|~9|~10",
+        "illuminate/support": "~8|~9|~10",
         "phayes/geophp": "~1.2"
     },
     "require-dev": {
-        "mockery/mockery": "^1.3.1",
-        "phpunit/phpunit": "^8.4|^9.0",
-        "psy/psysh": "^0.10",
-        "symfony/thanks": "^1.1",
-        "symfony/var-dumper": "^5.0"
+        "laravel/pint": "^1.6",
+        "mockery/mockery": "^1.5.1",
+        "phpunit/phpunit": "^9.6.5",
+        "psy/psysh": "^0.11.1",
+        "symfony/var-dumper": "^6.2"
     },
     "autoload": {
         "psr-4": {
@@ -54,9 +54,9 @@
     },
     "config": {
         "sort-packages": true,
-	"allow-plugins": {
-	    "symfony/thanks": true
-	}
+        "allow-plugins": {
+            "symfony/thanks": false
+        }
     },
     "minimum-stability": "dev",
     "prefer-stable": true
