@@ -8,28 +8,24 @@ use InvalidArgumentException;
  * Class TypeMapper
  *
  * Convert the types that geoPHP supports in more predictable names.
- *
- * @package Spinen\Geometry\Support
  */
 class TypeMapper
 {
     /**
      * Supported geometry types.
-     *
-     * @var array
      */
-    protected $types = [
-        'Ewkb'          => 'ewkb',
-        'Ewkt'          => 'ewkt',
-        'GeoHash'       => 'geohash',
-        'GeoJson'       => 'geojson',
-        'GeoRss'        => 'georss',
+    protected array $types = [
+        'Ewkb' => 'ewkb',
+        'Ewkt' => 'ewkt',
+        'GeoHash' => 'geohash',
+        'GeoJson' => 'geojson',
+        'GeoRss' => 'georss',
         'GoogleGeocode' => 'google_geocode',
-        'Gpx'           => 'gpx',
-        'Json'          => 'json',
-        'Kml'           => 'kml',
-        'Wkb'           => 'wkb',
-        'Wkt'           => 'wkt',
+        'Gpx' => 'gpx',
+        'Json' => 'json',
+        'Kml' => 'kml',
+        'Wkb' => 'wkb',
+        'Wkt' => 'wkt',
     ];
 
     /**
@@ -37,17 +33,14 @@ class TypeMapper
      *
      * Look it up in the types to make sure that it is defined & map it to the string that geoPHP expects.
      *
-     * @param string $type
-     *
-     * @return string
      * @throws InvalidArgumentException
      */
-    public function map($type)
+    public function map(string $type): string
     {
         if (in_array($type, array_keys($this->types))) {
             return $this->types[$type];
         }
 
-        throw new InvalidArgumentException(sprintf("Unknown geometry type of [%s] was provided.", $type));
+        throw new InvalidArgumentException(sprintf('Unknown geometry type of [%s] was provided.', $type));
     }
 }
