@@ -11,8 +11,6 @@ use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 
 /**
  * Class TestCase
- *
- * @package Tests\Spinen\BrowserFilter
  */
 abstract class TestCase extends PHPUnitTestCase
 {
@@ -22,11 +20,6 @@ abstract class TestCase extends PHPUnitTestCase
      * Helper to allow mocking of iterator classes.
      *
      * @link https://gist.github.com/VladaHejda/8299871
-     *
-     * @param MockInterface $mock
-     * @param array $items
-     *
-     * @return void
      */
     protected function mockArrayIterator(MockInterface $mock, array $items): void
     {
@@ -49,7 +42,7 @@ abstract class TestCase extends PHPUnitTestCase
             $counter = 0;
 
             $mock->shouldReceive('rewind')
-                 ->andReturnUsing(function () use (& $counter) {
+                 ->andReturnUsing(function () use (&$counter) {
                      $counter = 0;
                  });
 
@@ -57,23 +50,23 @@ abstract class TestCase extends PHPUnitTestCase
             $keys = array_values(array_keys($items));
 
             $mock->shouldReceive('valid')
-                 ->andReturnUsing(function () use (& $counter, $vals) {
+                 ->andReturnUsing(function () use (&$counter, $vals) {
                      return isset($vals[$counter]);
                  });
 
             $mock->shouldReceive('current')
-                 ->andReturnUsing(function () use (& $counter, $vals) {
+                 ->andReturnUsing(function () use (&$counter, $vals) {
                      return $vals[$counter];
                  });
 
             $mock->shouldReceive('key')
-                 ->andReturnUsing(function () use (& $counter, $keys) {
+                 ->andReturnUsing(function () use (&$counter, $keys) {
                      return $keys[$counter];
                  });
 
             $mock->shouldReceive('next')
-                 ->andReturnUsing(function () use (& $counter) {
-                     ++$counter;
+                 ->andReturnUsing(function () use (&$counter) {
+                     $counter++;
                  });
         }
 
